Let `ScanResultsStorage` configure itself
Move the code to configure the scan results storage from the
`ScannerCommand` directly to `ScanResultsStorage`. This allows to use
the same configuration in other places.

With this change the default storage is only created if no other
storage is configured. Accordingly, the number of scan result files in
the default storage is also only logged, if the default storage is used.
